Add in login Page and Post Feed Page
login username will be the User.currentuser to show name for all later post and messages. Usernames are not saved.

Renamed commentButton to PinkButton because will be used on other buttons rather than just the comment. So also made function to configure text of button.

Add in tableView and pink button to Post Feed