Add RadioButton that correspond to Offer or Request postType
This is the component for the MakeEventVC toggle that will be added later.
If one is currently editing a post of type request, the request radio button will be active while the offer radio button will be inactive.
Likewise in the case that offer is the one active, the request radio button will be inactive.
This is just the setup (something in MakeEventVC will have to configure it later)