Connected PostForm to MakePostVC
This setups postform base on Post Type (request or offer). But more importantly (since the attributes for request and offer are identical), it manages the toggle bar up on the top, making the radiobutton light up according to its Post Type.

One can almost create a post. Network for post still needs to be setup later when off the plane.