Set up the tableview in Item's Post VC with corresponding posts from Item
There will be however many cell as item.posts.count
Each of the post in the item will configure the cell.
Cell Reuse is used.
When the cell is clicked on, one goes to the chat page for that post (push)
If create Post button is clicked, now it goes to make post controller (present).